requirements.txt
requests-2.25.1
moviepy-1.0.3
eyed3-0.9.6